Refactor test details query generators as well
